Add SeDebugPrivilege helper function
